fix : 당첨금이 너무 커질경우 int의 범위를 넘어가는 문제 수정
당첨금을 담는 변수의 타입을 int -> long 으로 변경

+++ src/main/java/lotto/domain/LottoResultStatistics.java
@@ -33,8 +33,8 @@ public class LottoResultStatistics {
         return lottoResult;
     }
 
-    public int calculateEarning(final Money money) {
-        int totalReward = Arrays.stream(LottoRank.values())
+    public long calculateEarning(final Money money) {
+        long totalReward = Arrays.stream(LottoRank.values())
                 .mapToInt(lottoRank -> this.lottoResult.get(lottoRank) * lottoRank.getReward())
                 .sum();
